Refactors query types for delete and find
Updates the `delete` and `find` methods in the `Collection` class to use a more specific type for the query parameter, improving type safety and consistency.

Removes the need to define the query type inline and uses `Partial<Doc<S>>` instead.

Also adds a line to index.ts to test the delete functionality.